Create README.md
I assume we aren't moving into the house tomorrow as you literally have not said a word to me in the past to weeks and there is a family that lives there. My stuff is being delivered to a public storage place tomorrow morning. Would really appreciate to know where we're supposed to live Patrice